requirements.txt
antlr4-python3-runtime  > 4.5
lxml >= 3.5.0
